Adds an email-form class to columns.
With the class "email-form" in the columns block we will build out a short form based off the link in the block that will direct users to the page in the link and pass in the link text as the CTA.

Form validation for empty form as well as incorrect email.

Email is passed into the landing page header and will be read and passed into the field.